requirements/base.txt
Django==2.0.7
psycopg2-binary==2.7.5
pytz==2018.5
Pillow==5.2.0
django-extensions==2.1.0
djangorestframework==3.8.2
django-widget-tweaks==1.4.2
django-axes==4.4.0
django-passwords==0.3.12
django-autocomplete-light==3.2.10
django-querysetsequence==0.8
awesome-slugify==1.6.5
django-el-pagination==3.2.4
python-decouple==3.1
django-tables2-bootstrap4==2.0.0b5
sorl-thumbnail==12.4.1
phonenumbers==8.9.10
django-phonenumber-field==2.0.0
django-crispy-forms==1.7.2
flake8==3.5.0
